Internal: relocate utilities that should not be public
