+++ src/_utils/beautify-html/__tests__/index.js
@@ -1,0 +1,17 @@
+import test from "ava"; import "babel-core/register"
+
+import beautifyHTML from ".."
+
+test("should make html readable", (t) => (
+  t.is(
+    beautifyHTML("<html><body><h1>title</h1><p>test</p></body></html>"),
+    `<html>
+
+<body>
+  <h1>title</h1>
+  <p>test</p>
+</body>
+
+</html>`
+  )
+))
