+++ src/md-collection-loader/index.js
@@ -34,8 +34,8 @@ import loaderUtils from "loader-utils"
 import yamlHeaderParser from "gray-matter"
 import markdownIt from "markdown-it"
 
-import joinUri from "../_utils/join-uri.js"
-import toUri from "../to-uri"
+import joinUri from "../_utils/join-uri"
+import toUri from "../_utils/to-uri"
 import enhanceCollection from "../enhance-collection"
 import feed from "./feed"
 import cache from "./cache"
